Se agrega guardado de cliente

+++ Model/Sys_Client.cs
@@ -14,7 +14,7 @@ namespace Model
     
     public partial class Sys_Client
     {
-        public int Sys_Id { get; set; }
+        public long Sys_Id { get; set; }
         public long Sys_IdCompany { get; set; }
         public string Sys_Rfc { get; set; }
         public string Sys_BusinessName { get; set; }
@@ -28,7 +28,7 @@ namespace Model
         public string Sys_Country { get; set; }
         public System.DateTime Sys_RegistrationDate { get; set; }
         public System.DateTime Sys_ModificationDate { get; set; }
-        public int Sys_UserId { get; set; }
+        public long Sys_UserId { get; set; }
         public string Sys_RfcCompany { get; set; }
     }
 }
